Make addons configurable.
The default set of addons can be found in DEFAULT_ADDONS. However a
custom set of addons can be configured with:

<configuration>
    <addons>
        <addon>
            <groupId>mygroup</groupId>
            <artifactId>myartifact</artifactId>
            <classifier>myclassifier</classifier>
        </addon>
    </addons>
</configuration>

Addons are found by looking for a dependency in the current project or
in the dependency management that matches the groupId and the artifactID
of the specified addons. If one is found, the version that that dependency
has is used for the addon feature model. The addon feature model is
assumed to have the specified classifier in its coordinates and be of type
'slingosgifeature'.